Move tests out of lib
